refactor(ui): split media_card.dart into focused modules
Break the ~1000-line MediaCard primitive into helpers/badges/tile/row under media_card/, with a barrel preserving the public import path.
